removing mistakes

+++ ft_strncmp.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strncmp.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bschaafs <bschaafs@student.42.fr>          +#+  +:+       +#+        */
+/*   By: bootjan <bootjan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/11 22:32:59 by bootjan           #+#    #+#             */
-/*   Updated: 2023/10/10 15:16:13 by bschaafs         ###   ########.fr       */
+/*   Updated: 2023/10/10 16:46:06 by bootjan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,9 +19,9 @@ int	ft_strncmp(const char *s1, const char *s2, size_t n)
 	i = 0;
 	while ((s1[i] || s2[i]) && i < n)
 	{
-		if (s1[i] > s2[i])
+		if ((unsigned long)s1[i] > (unsigned long)s2[i])
 			return (1);
-		if (s1[i] < s2[i])
+		if ((unsigned long)s1[i] < (unsigned long)s2[i])
 			return (-1);
 		i++;
 	}
